requirements.txt
django==1.10.6
behave==1.2.5
django_behave==0.1.5
splinter==0.7.5
factory_boy==2.8.1
social-auth-app-django==1.1.0
selenium==3.0.2
